refactor to account for ancestry correctly. Crashes again :'(
We don't always run the BooNE models due to the KE check. Move to the NuBeamSteppingAction which always runs.
Use of smart pointers ensures we get trajectory points and do not skip any tracks, so ancestry is happy.
But there's an extra deletion used somewhere when we clear the hashmap, suggesting shared ptr...